Fix bug VIP-156

+++ module/vipshop-admin/QA/type.go
@@ -46,7 +46,7 @@ type TextData struct {
 type CUData struct {
 	Type     string        `json:"type"`
 	Value    string        `json:"value"`
-	Score    int           `json:"score"`
+	Score    float64       `json:"score"`
 	Category string        `json:"category"`
 	Data     []interface{} `json:"data"`
 }
